\name{JamachineLearning-package}
\alias{JamachineLearning-package}
\alias{JamachineLearning}
\docType{package}
\title{
\packageTitle{JamachineLearning}
}
\description{
\packageDescription{JamachineLearning}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{JamachineLearning}
\packageIndices{JamachineLearning}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{JamachineLearning}

Maintainer: \packageMaintainer{JamachineLearning}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
