% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_0.R
\name{loadJML}
\alias{loadJML}
\title{Load training and test datasets}
\usage{
loadJML(training, testing)
}
\arguments{
\item{training}{path to training csv.}

\item{testing}{path to test csv.}
}
\value{
list with 2 data.tables 'training' and 'testing'
\item{training}{training data.table}
\item{testing}{testing data.table}
}
\description{
This function loads both csv's corresponding to the training and test dataset.
}
\details{
Each of the two datasets are loaded from csv's as data.tables. Column "user_name" is defined as key in each. Variable "cvtd_timestamp" is converted to class POSIXct. All relevant variables are classed as numericals, interpreting emtpy cells "" and "#DIV/0!" as NA.
}
\examples{
# load with folder csv's in working directory containing both files.
# dlist <- loadJML('./csvs/pml-testing.csv', './csvs/pml-testing.csv')
}

